Moving BufferedChecksum classes back to main
