+++ server/src/main/java/org/opensearch/index/translog/BufferedChecksumStreamOutput.java
@@ -30,10 +30,11 @@
  * GitHub history for details.
  */
 
-package org.opensearch.core.common.io.stream;
+package org.opensearch.index.translog;
 
 import org.apache.lucene.store.BufferedChecksum;
 import org.opensearch.common.annotation.PublicApi;
+import org.opensearch.core.common.io.stream.StreamOutput;
 
 import java.io.IOException;
 import java.util.zip.CRC32;
